requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
pyyaml
celery
redis
aiofiles